[modify]move directory.
