+++ example/components/snabbdom/counter.js
