+++ example/components/snabbdom/decrementer.js
